[add] 회원탈퇴 다이얼로그에 회원탈퇴 요청 추가

+++ libs/components/src/lib/AccountRemoveDialog.tsx
@@ -10,26 +10,49 @@ import {
   ModalProps,
   Text,
   Input,
+  useToast,
 } from '@chakra-ui/react';
-import { useProfile } from '@project-lc/hooks';
+import { useDeleteSellerMutation, useProfile } from '@project-lc/hooks';
 import { useForm } from 'react-hook-form';
 
 export type AccountRemoveDialogProps = Pick<ModalProps, 'isOpen' | 'onClose'> & {
+  /** remove 요청 성공 후 콜백 */
   onRemove: () => void;
 };
 
 export function AccountRemoveDialog(props: AccountRemoveDialogProps): JSX.Element {
   const { isOpen, onClose, onRemove } = props;
   const { data } = useProfile();
-  const { register, handleSubmit, watch, reset } = useForm();
+  const { register, handleSubmit, watch, reset, getValues } = useForm();
+  const { mutateAsync } = useDeleteSellerMutation();
+  const toast = useToast();
 
   const closeModal = () => {
     reset();
     onClose();
   };
+
   const deleteAccount = () => {
-    onRemove();
-    closeModal();
+    mutateAsync(getValues('email'))
+      .then((isDeleted) => {
+        if (isDeleted) {
+          onRemove();
+          closeModal();
+        }
+      })
+      .catch((error) => {
+        if (error.response.status === 400) {
+          toast({
+            title: '탈퇴 오류',
+            description: error.response.data.message,
+            status: 'error',
+          });
+        }
+        toast({
+          title: '탈퇴 오류',
+          status: 'error',
+        });
+      });
   };
   return (
     <Modal isOpen={isOpen} onClose={closeModal} isCentered>
@@ -48,18 +71,18 @@ export function AccountRemoveDialog(props: AccountRemoveDialogProps): JSX.Elemen
           <Text mb={3}>
             연결된 소셜 계정이 있다면 소셜 계정 연결 해제를 먼저 진행해주세요.
           </Text>
-          <Text>계속 진행하려면 이메일을 입력해주세요</Text>
+          <Text mb={2}>계속 진행하려면 로그인 한 계정의 이메일을 입력해주세요</Text>
           <Input {...register('email', { required: true })} />
         </ModalBody>
 
         <ModalFooter>
-          <Button colorScheme="blue" mr={3} onClick={closeModal}>
+          <Button mr={3} onClick={closeModal}>
             취소
           </Button>
           <Button
             colorScheme="red"
             type="submit"
-            disabled={!data || (data && data.sub !== watch('email'))}
+            // disabled={!data || (data && data.sub !== watch('email'))}
           >
             회원 탈퇴
           </Button>
